+++ galsim/compound.py
@@ -70,21 +70,32 @@ class Add(GSObject):
         elif len(args) == 1:
             # 1 argument.  Should be either a GSObject or a list of GSObjects
             if isinstance(args[0], GSObject):
-                SBList = [args[0].SBProfile]
+                args = [args[0]]
             elif isinstance(args[0], list):
-                SBList = []
-                for obj in args[0]:
-                    if isinstance(obj, GSObject):
-                        SBList.append(obj.SBProfile)
-                    else:
-                        raise TypeError("Input list must contain only GSObjects.")
+                args = args[0]
             else:
                 raise TypeError("Single input argument must be a GSObject or list of them.")
-            GSObject.__init__(self, galsim.SBAdd(SBList, gsparams=gsparams))
-        elif len(args) >= 2:
-            # >= 2 arguments.  Convert to a list of SBProfiles
+        # else args is already the list of objects
+
+        if len(args) == 1:
+            # No need to make an SBAdd in this case.
+            GSObject.__init__(self, args[0])
+            if hasattr(args[0],'noise'): 
+                self.noise = args[0].noise
+        else:
+            # If any of the objects have a noise attribute, then we propagate the sum of the 
+            # noises (they add like variances) to the final sum.
+            noise = None
+            for obj in args:
+                if hasattr(obj,'noise'):
+                    if noise is None: 
+                        noise = obj.noise
+                    else: 
+                        noise += obj.noise
             SBList = [obj.SBProfile for obj in args]
             GSObject.__init__(self, galsim.SBAdd(SBList, gsparams=gsparams))
+            if noise is not None: 
+                self.noise = noise
 
 class Convolve(GSObject):
     """A class for convolving 2 or more GSObjects.
@@ -118,6 +129,11 @@ class Convolve(GSObject):
     Also, note that parameters related to the Fourier-space calculations must be set when
     initializing the individual GSObjects that go into the Convolve, NOT when creating the Convolve
     (at which point the accuracy and threshold parameters will simply be ignored).
+
+    @param args       Unnamed args should be a list of objects to convolve.
+    @param real_space Whether to use real space convolution.  Default is to automatically select
+                      this according to whether the object has hard edges.
+    @param gsparams   Optional gsparams argument.
     """
                     
     # --- Public Class methods ---
@@ -130,27 +146,19 @@ class Convolve(GSObject):
             raise ValueError("Convolve must be initialized with at least one GSObject.")
         elif len(args) == 1:
             if isinstance(args[0], GSObject):
-                SBList = [args[0].SBProfile]
+                args = [args[0]]
             elif isinstance(args[0], list):
-                SBList=[]
-                for obj in args[0]:
-                    if isinstance(obj, GSObject):
-                        SBList.append(obj.SBProfile)
-                    else:
-                        raise TypeError("Input list must contain only GSObjects.")
+                args = args[0]
             else:
                 raise TypeError("Single input argument must be a GSObject or list of them.")
-        elif len(args) >= 2:
-            # >= 2 arguments.  Convert to a list of SBProfiles
-            SBList = []
-            for obj in args:
-                if isinstance(obj, GSObject):
-                    SBList.append(obj.SBProfile)
-                else:
-                    raise TypeError("Input args must contain only GSObjects.")
+        # else args is already the list of objects
 
-        # Having built the list of SBProfiles or thrown exceptions if necessary, see now whether
-        # to perform real space convolution...
+        if len(args) == 1:
+            # No need to make an SBConvolve in this case.  Can early exit.
+            GSObject.__init__(self, args[0])
+            if hasattr(args[0],'noise'): 
+                self.noise = args[0].noise
+            return
 
         # Check kwargs
         # real_space can be True or False (default if omitted is None), which specifies whether to 
@@ -159,7 +167,6 @@ class Convolve(GSObject):
         # we will usually do the fourier method.  However, if there are 2 components _and_ both of 
         # them have hard edges, then we use real-space convolution.
         real_space = kwargs.pop("real_space", None)
-
         gsparams = kwargs.pop("gsparams", None)
 
         # Make sure there is nothing left in the dict.
@@ -167,22 +174,17 @@ class Convolve(GSObject):
             raise TypeError(
                 "Convolve constructor got unexpected keyword argument(s): %s"%kwargs.keys())
 
-
-        # If 1 argument, check if it is a list:
-        if len(args) == 1 and isinstance(args[0], list):
-            args = args[0]
-
+        # Check whether to perform real space convolution...
+        # Start by checking if all objects have a hard edge.
         hard_edge = True
         for obj in args:
             if not obj.hasHardEdges():
                 hard_edge = False
 
         if real_space is None:
-            # Figure out if it makes more sense to use real-space convolution.
+            # The automatic determination is to use real_space if 2 items, both with hard edges.
             if len(args) == 2:
                 real_space = hard_edge
-            elif len(args) == 1:
-                real_space = obj.isAnalyticX()
             else:
                 real_space = False
         
@@ -222,9 +224,32 @@ class Convolve(GSObject):
                         real_space = False
                         break
 
-        # Then finally initialize the SBProfile using the objects' SBProfiles in SBList
+        # If one of the objects has a noise attribute, then we convolve it by the others.
+        # More than one is not allowed.
+        noise = None
+        noise_convolve = []
+        for obj in args:
+            if hasattr(obj,'noise'):
+                if noise is not None: 
+                    import warnings
+                    warnings.warn("Unable to propagate noise in galsim.Convolve when multiple "+
+                                  "objects have noise attribute")
+                    noise = None
+                    break
+                noise = obj.noise
+                others = [ obj2 for obj2 in args if obj2 is not obj ]
+                assert len(others) > 0
+                if len(others) == 1: 
+                    noise.convolveWith(others[0])
+                else: 
+                    noise.convolveWith(galsim.Convolve(others))
+
+        # Then finally initialize the SBProfile using the objects' SBProfiles.
+        SBList = [ obj.SBProfile for obj in args ]
         GSObject.__init__(self, galsim.SBConvolve(SBList, real_space=real_space,
                                                   gsparams=gsparams))
+        if noise is not None: 
+            self.noise = noise
 
 
 class Deconvolve(GSObject):
@@ -238,14 +263,18 @@ class Deconvolve(GSObject):
     help(galsim.GSParams) for more information about this option.  Note: if gsparams is unspecified
     (or None), then the Deconvolve instance inherits the same GSParams as the object being
     deconvolved.
+
+    @param obj        The object to deconvolve.
+    @param gsparams   Optional gsparams argument.
     """
     # --- Public Class methods ---
-    def __init__(self, farg, gsparams=None):
-        if isinstance(farg, GSObject):
-            self.farg = farg
-            GSObject.__init__(self, galsim.SBDeconvolve(self.farg.SBProfile, gsparams=gsparams))
-        else:
+    def __init__(self, obj, gsparams=None):
+        if not isinstance(obj, GSObject):
             raise TypeError("Argument to Deconvolve must be a GSObject.")
+        GSObject.__init__(self, galsim.SBDeconvolve(obj.SBProfile, gsparams=gsparams))
+        if hasattr(obj,'noise'):
+            import warnings
+            warnings.warn("Unable to propagate noise in galsim.Deconvolve")
 
 
 class AutoConvolve(GSObject):
@@ -254,19 +283,47 @@ class AutoConvolve(GSObject):
     It is equivalent in functionality to galsim.Convolve([obj,obj]), but takes advantage of
     the fact that the two profiles are the same for some efficiency gains.
 
-    @param obj       The object to be convolved with itself.
-    @param gsparams  You may also specify a gsparams argument.  See the docstring for
-                     galsim.GSParams using help(galsim.GSParams) for more information about this
-                     option.  Note that parameters related to the Fourier-space calculations must be
-                     set when initializing the GSObject that goes into the AutoConvolve, NOT when
-                     creating the AutoConvolve (at which point the accuracy and threshold parameters
-                     will simply be ignored).
+    @param obj        The object to be convolved with itself.
+    @param real_space Whether to use real space convolution.  Default is to automatically select
+                      this according to whether the object has hard edges.
+    @param gsparams   Optional gsparams argument.
     """
     # --- Public Class methods ---
-    def __init__(self, obj, gsparams=None):
+    def __init__(self, obj, real_space=None, gsparams=None):
         if not isinstance(obj, GSObject):
             raise TypeError("Argument to AutoConvolve must be a GSObject.")
-        GSObject.__init__(self, galsim.SBAutoConvolve(obj.SBProfile, gsparams=gsparams))
+
+        # Check whether to perform real space convolution...
+        # Start by checking if obj has a hard edge.
+        hard_edge = obj.hasHardEdges()
+
+        if real_space is None:
+            # The automatic determination is to use real_space if obj has hard edges.
+            real_space = hard_edge
+        
+        # Warn if doing DFT convolution for objects with hard edges.
+        if not real_space and hard_edge:
+            import warnings
+            msg = """
+            Doing auto-convolution of object with hard edges.
+            This might be more accurate and/or faster using real_space=True"""
+            warnings.warn(msg)
+
+        # Can't do real space if object is not analytic, so check for that.
+        if real_space and not obj.isAnalyticX():
+            import warnings
+            msg = """
+            Object to be auto-convolved is not analytic in real space.
+            Cannot use real space convolution.
+            Switching to DFT method."""
+            warnings.warn(msg)
+            real_space = False
+
+        GSObject.__init__(self, galsim.SBAutoConvolve(obj.SBProfile, real_space=real_space,
+                                                      gsparams=gsparams))
+        if hasattr(obj,'noise'):
+            import warnings
+            warnings.warn("Unable to propagate noise in galsim.AutoConvolve")
 
 
 class AutoCorrelate(GSObject):
@@ -279,20 +336,48 @@ class AutoCorrelate(GSObject):
     This class is primarily targeted for use by the galsim.CorrelatedNoise models when convolving 
     with a GSObject.
 
-    @param obj       The object to be correlated with itself.
-
-    @param gsparams  You may also specify a gsparams argument.  See the docstring for
-                     galsim.GSParams using help(galsim.GSParams) for more information about this
-                     option.  Note that parameters related to the Fourier-space calculations must be
-                     set when initializing the GSObject that goes into the AutoCorrelate, NOT when
-                     creating the AutoCorrelate (at which point the accuracy and threshold
-                     parameters will simply be ignored).
+    @param obj        The object to be convolved with itself.
+    @param real_space Whether to use real space convolution.  Default is to automatically select
+                      this according to whether the object has hard edges.
+    @param gsparams   Optional gsparams argument.
     """
     # --- Public Class methods ---
-    def __init__(self, obj, gsparams=None):
+    def __init__(self, obj, real_space=None, gsparams=None):
         if not isinstance(obj, GSObject):
             raise TypeError("Argument to AutoCorrelate must be a GSObject.")
-        GSObject.__init__(self, galsim.SBAutoCorrelate(obj.SBProfile, gsparams=gsparams))
+
+        # Check whether to perform real space convolution...
+        # Start by checking if obj has a hard edge.
+        hard_edge = obj.hasHardEdges()
+
+        if real_space is None:
+            # The automatic determination is to use real_space if obj has hard edges.
+            real_space = hard_edge
+        
+        # Warn if doing DFT convolution for objects with hard edges.
+        if not real_space and hard_edge:
+            import warnings
+            msg = """
+            Doing auto-convolution of object with hard edges.
+            This might be more accurate and/or faster using real_space=True"""
+            warnings.warn(msg)
+
+        # Can't do real space if object is not analytic, so check for that.
+        if real_space and not obj.isAnalyticX():
+            import warnings
+            msg = """
+            Object to be auto-convolved is not analytic in real space.
+            Cannot use real space convolution.
+            Switching to DFT method."""
+            warnings.warn(msg)
+            real_space = False
+
+        GSObject.__init__(self, galsim.SBAutoCorrelate(obj.SBProfile, real_space=real_space,
+                                                       gsparams=gsparams))
+
+        if hasattr(obj,'noise'):
+            import warnings
+            warnings.warn("Unable to propagate noise in galsim.AutoCorrelate")