+++ src/SBInterpolatedImage.cpp
@@ -40,10 +40,9 @@ namespace galsim {
     SBInterpolatedImage::SBInterpolatedImage(
         const BaseImage<T>& image,
         boost::shared_ptr<Interpolant2d> xInterp, boost::shared_ptr<Interpolant2d> kInterp,
-        double dx, double pad_factor, boost::shared_ptr<BaseImage<T> > pad_image,
-        const GSParamsPtr& gsparams) :
+        double dx, double pad_factor, const GSParamsPtr& gsparams) :
         SBProfile(new SBInterpolatedImageImpl(
-                image,xInterp,kInterp,dx,pad_factor,pad_image,gsparams)) {}
+                image,xInterp,kInterp,dx,pad_factor,gsparams)) {}
 
     SBInterpolatedImage::SBInterpolatedImage(
         const MultipleImageHelper& multi, const std::vector<double>& weights,
@@ -55,16 +54,16 @@ namespace galsim {
 
     SBInterpolatedImage::~SBInterpolatedImage() {}
 
-    void SBInterpolatedImage::calculateStepK() const 
+    void SBInterpolatedImage::calculateStepK(double max_stepk) const 
     { 
         assert(dynamic_cast<const SBInterpolatedImageImpl*>(_pimpl.get()));
-        return static_cast<const SBInterpolatedImageImpl&>(*_pimpl).calculateStepK(); 
+        return static_cast<const SBInterpolatedImageImpl&>(*_pimpl).calculateStepK(max_stepk); 
     }
 
-    void SBInterpolatedImage::calculateMaxK() const 
+    void SBInterpolatedImage::calculateMaxK(double max_maxk) const 
     {
         assert(dynamic_cast<const SBInterpolatedImageImpl*>(_pimpl.get()));
-        return static_cast<const SBInterpolatedImageImpl&>(*_pimpl).calculateMaxK(); 
+        return static_cast<const SBInterpolatedImageImpl&>(*_pimpl).calculateMaxK(max_maxk); 
     }
 
     template <class T>
@@ -147,8 +146,7 @@ namespace galsim {
 
     template <class T>
     MultipleImageHelper::MultipleImageHelper(const BaseImage<T>& image, double dx, 
-        double pad_factor, boost::shared_ptr<BaseImage<T> > pad_image) :
-        _pimpl(new MultipleImageHelperImpl)
+        double pad_factor) : _pimpl(new MultipleImageHelperImpl)
     {
         dbg<<"Start MultipleImageHelper constructor for one image\n";
         dbg<<"image bounds = "<<image.getBounds()<<std::endl;
@@ -168,14 +166,6 @@ namespace galsim {
         if (pad_factor <= 0.) pad_factor = sbp::default_pad_factor;
         _pimpl->Nk = goodFFTSize(int(pad_factor*_pimpl->Ninitial));
 
-        // Make sure we can fit the pad_image.
-        if (pad_image) {
-            int pad_size = std::max( pad_image->getYMax()-pad_image->getYMin()+1,
-                                     pad_image->getXMax()-pad_image->getXMin()+1 );
-            dbg<<"pad_size = "<<pad_size<<std::endl;
-            if (pad_size > _pimpl->Nk) _pimpl->Nk = pad_size;
-        }
-
         dbg<<"Ninitial = "<<_pimpl->Ninitial<<std::endl;
         dbg<<"Nk = "<<_pimpl->Nk<<std::endl;
 
@@ -189,20 +179,6 @@ namespace galsim {
         _pimpl->xflux.resize(1);
         _pimpl->yflux.resize(1);
         _pimpl->vx[0].reset(new XTable(_pimpl->Nk, _pimpl->dx));
-
-        if (pad_image.get()) {
-            // Start by copying the pad_image
-            dbg<<"Copying pad_image\n";
-            int xStart = -((pad_image->getXMax()-pad_image->getXMin()+1)/2);
-            int y = -((pad_image->getYMax()-pad_image->getYMin()+1)/2);
-            dbg<<"xStart = "<<xStart<<", yStart = "<<y<<std::endl;
-            for (int iy = pad_image->getYMin(); iy<=pad_image->getYMax(); ++iy, ++y) {
-                int x = xStart;
-                for (int ix = pad_image->getXMin(); ix<=pad_image->getXMax(); ++ix, ++x) {
-                    _pimpl->vx[0]->xSet(x, y, (*pad_image)(ix,iy));
-                }
-            }
-        }
 
         // Copy the given image to the center
         // Also accumulate the flux and centroid of the original image
@@ -242,10 +218,9 @@ namespace galsim {
     SBInterpolatedImage::SBInterpolatedImageImpl::SBInterpolatedImageImpl(
         const BaseImage<T>& image, 
         boost::shared_ptr<Interpolant2d> xInterp, boost::shared_ptr<Interpolant2d> kInterp,
-        double dx, double pad_factor, boost::shared_ptr<BaseImage<T> > pad_image,
-        const GSParamsPtr& gsparams) :
+        double dx, double pad_factor, const GSParamsPtr& gsparams) :
         SBProfileImpl(gsparams),
-        _multi(image,dx,pad_factor,pad_image), _wts(1,1.), _xInterp(xInterp), _kInterp(kInterp),
+        _multi(image,dx,pad_factor), _wts(1,1.), _xInterp(xInterp), _kInterp(kInterp),
         _readyToShoot(false)
     { initialize(); }
 
@@ -611,11 +586,12 @@ namespace galsim {
     // size region around the center encloses (1-alias_threshold) of the total flux.
     // This can be useful if you make the image bigger than you need to, just to be
     // safe, but then want to use as large a stepk value as possible.
-    void SBInterpolatedImage::SBInterpolatedImageImpl::calculateStepK() const
+    void SBInterpolatedImage::SBInterpolatedImageImpl::calculateStepK(double max_stepk) const
     {
         dbg<<"Start SBInterpolatedImage calculateStepK()\n";
         dbg<<"Current value of stepk = "<<_stepk<<std::endl;
         dbg<<"Find box that encloses "<<1.-this->gsparams->alias_threshold<<" of the flux.\n";
+        dbg<<"Max_stepk = "<<max_stepk<<std::endl;
         dbg<<"xtab size = "<<_xtab->getN()<<", scale = "<<_xtab->getDx()<<std::endl;
         //int N = _xtab->getN();
         double scale = _xtab->getDx();
@@ -639,6 +615,9 @@ namespace galsim {
         int dy = b.getYMin() + ((b.getYMax()-b.getYMin()+1)/2);
         dbg<<"b = "<<b<<std::endl;
         dbg<<"dx,dy = "<<dx<<','<<dy<<std::endl;
+        int min_d = max_stepk == 0. ? 0 : int(ceil(M_PI/max_stepk/scale));
+        dbg<<"min_d = "<<min_d<<std::endl;
+        double max_flux = flux;
         for (int d=1; d<=Nino2; ++d) {
             xdbg<<"d = "<<d<<std::endl;
             xdbg<<"d1 = "<<d1<<std::endl;
@@ -652,15 +631,26 @@ namespace galsim {
                 if (b.includes(Position<int>(-x+dx,d+dy))) flux += _xtab->xval(-x,d);  // top
                 if (b.includes(Position<int>(-d+dx,-x+dy))) flux += _xtab->xval(-d,-x); // left
             }
+            if (flux > max_flux) {
+                max_flux = flux;
+                if (flux > fluxTot) {
+                    // If flux w/in some radius is more than the total, then we have a case of
+                    // noise artificially lowering the nominal flux.  We will use the radius
+                    // of the maximum flux we get during this procedure.
+                    d1 = d;
+                }
+            }
             if (flux < thresh) {
                 d1 = 0; // Mark that we haven't gotten to a good enclosing radius yet.
-            } else {
+            } else if (d > min_d) {
                 if (d1 == 0) d1 = d; // Mark this radius as a good one.
             }
         }
         dbg<<"Done: flux = "<<flux<<std::endl;
+        dbg<<"max_flux = "<<max_flux<<", current fluxTot = "<<fluxTot<<std::endl;
         // Should have added up to the total flux.
         assert( std::abs(flux - fluxTot) < 1.e-3 * std::abs(fluxTot) );
+
         if (d1 == 0) {
             dbg<<"No smaller radius found.  Keep current value of stepk\n";
             return;
@@ -684,10 +674,11 @@ namespace galsim {
     inline double fast_norm(const std::complex<double>& z)
     { return real(z)*real(z) + imag(z)*imag(z); }
 
-    void SBInterpolatedImage::SBInterpolatedImageImpl::calculateMaxK() const
+    void SBInterpolatedImage::SBInterpolatedImageImpl::calculateMaxK(double max_maxk) const
     {
         dbg<<"Start SBInterpolatedImage calculateMaxK()\n";
         dbg<<"Current value of maxk = "<<_maxk<<std::endl;
+        dbg<<"max_maxk = "<<max_maxk<<std::endl;
         dbg<<"Find the smallest k such that all values outside of this are less than "
             <<this->gsparams->maxk_threshold<<std::endl;
         checkK();
@@ -703,14 +694,16 @@ namespace galsim {
         int n_below_thresh = 0;
         int N = _ktab->getN();
         // Don't go past the current value of maxk
-        if (N/2 * dk > _maxk) 
-            N = int(_maxk*2./dk);
+        if (max_maxk == 0.) max_maxk = _maxk;
+        int max_ix = int(std::ceil(max_maxk / dk));
+        if (max_ix > N/2) max_ix = N/2;
+
         // We take the k value to be maximum of kx and ky.  This is appropriate, because
         // this is how maxK() is eventually used -- it sets the size in k-space for both
         // kx and ky when drawing.  Since kx<0 is just the conjugate of the corresponding
         // point at (-kx,-ky), we only check the right half of the square.  i.e. the 
         // upper-right and lower-right quadrants.
-        for(int ix=0; ix<=N/2; ++ix) {
+        for(int ix=0; ix<=max_ix; ++ix) {
             xdbg<<"Start search for ix = "<<ix<<std::endl;
             // Search along the two sides with either kx = ix or ky = ix.
             for(int iy=0; iy<=ix; ++iy) {
@@ -861,22 +854,18 @@ namespace galsim {
     template SBInterpolatedImage::SBInterpolatedImage(
         const BaseImage<float>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<float> > pad_image,
         const GSParamsPtr& gsparams);
     template SBInterpolatedImage::SBInterpolatedImage(
         const BaseImage<double>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<double> > pad_image,
         const GSParamsPtr& gsparams);
     template SBInterpolatedImage::SBInterpolatedImage(
         const BaseImage<int32_t>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<int32_t> > pad_image,
         const GSParamsPtr& gsparams);
     template SBInterpolatedImage::SBInterpolatedImage(
         const BaseImage<int16_t>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<int16_t> > pad_image,
         const GSParamsPtr& gsparams);
 
     template MultipleImageHelper::MultipleImageHelper(
@@ -893,37 +882,29 @@ namespace galsim {
         double dx, double pad_factor);
 
     template MultipleImageHelper::MultipleImageHelper(
-        const BaseImage<float>& image, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<float> > pad_image);
+        const BaseImage<float>& image, double dx, double pad_factor);
     template MultipleImageHelper::MultipleImageHelper(
-        const BaseImage<double>& image, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<double> > pad_image);
+        const BaseImage<double>& image, double dx, double pad_factor);
     template MultipleImageHelper::MultipleImageHelper(
-        const BaseImage<int32_t>& image, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<int32_t> > pad_image);
+        const BaseImage<int32_t>& image, double dx, double pad_factor);
     template MultipleImageHelper::MultipleImageHelper(
-        const BaseImage<int16_t>& image, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<int16_t> > pad_image);
+        const BaseImage<int16_t>& image, double dx, double pad_factor);
 
     template SBInterpolatedImage::SBInterpolatedImageImpl::SBInterpolatedImageImpl(
         const BaseImage<float>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<float> > pad_image,
         const GSParamsPtr& gsparams);
     template SBInterpolatedImage::SBInterpolatedImageImpl::SBInterpolatedImageImpl(
         const BaseImage<double>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<double> > pad_image,
         const GSParamsPtr& gsparams);
     template SBInterpolatedImage::SBInterpolatedImageImpl::SBInterpolatedImageImpl(
         const BaseImage<int32_t>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<int32_t> > pad_image,
         const GSParamsPtr& gsparams);
     template SBInterpolatedImage::SBInterpolatedImageImpl::SBInterpolatedImageImpl(
         const BaseImage<int16_t>& image, boost::shared_ptr<Interpolant2d> xInterp,
         boost::shared_ptr<Interpolant2d> kInterp, double dx, double pad_factor,
-        boost::shared_ptr<BaseImage<int16_t> > pad_image,
         const GSParamsPtr& gsparams);
 
 } // namespace galsim