+++ CHANGELOG.md
@@ -9,6 +9,24 @@ Changes from v0.5 to current version:
 
 * Fixed a bug in the rendering of shifted images.  (Issue #424)
 
+* Made RealGalaxy objects keep track of their (correlated) noise.  Functions like applyShear,
+  Convolve, and such automatically keep this updated so the final object also knows the current
+  noise profile so it can be appropriately whitened.  (Issue #430)
+
+* Changed the noise padding options for RealGalaxy and InterpolatedImage somewhat.  Now there is
+  a parameter `noise_pad_size` which (optionally) sets a minimum size to pad out the image with
+  the same noise as is in the main image.  This is required to make sure whitening will work
+  correctly.  (Outside of this noise padding, `pad_factor` still sets the amount of zero-padding
+  to use to remove the ghost images in the FFT.)  (Issue #430)
+
+* Added whiten option to config for RealGalaxy objects to whiten the image.  (Issue #430)
+
+* Added VariableGaussianNoise to apply Gaussian noise with a variable sigma across the image.
+  (Issue #430)
+
+* Added `Current` type in config to use the current value of some other item in the config file
+  as part of a calculation.  Example of this is in demo11.yaml.  (Issue #430)
+
 * Added the offset parameter to the draw and drawShoot commands, and also to the constructor 
   of InterpolatedImage.  (Issue #439)
 
@@ -29,6 +47,10 @@ Changes from v0.5 to current version:
 * Added offset as a config option for the image field, which applies the given offset in pixels 
   when calling the draw command.  (Issue #449)
 
+* Fixed a bug in InterpolatedImage calculateStepK function for noisy images.  (Issue #454)
+
+* Added 'signal_to_noise' option in config for PSFs when there is not also a galaxy.  (Issue #459)
+
 * Added `galsim` executable to be preferred over either `galsim_yaml` or `galsim_json`, although 
   the old names are still valid for backwards compatibility.  (Issue #460)
 
@@ -38,3 +60,5 @@ Changes from v0.5 to current version:
   GalSim) module and use it with galsim.  Just use `galsim -m [your_module]` here.  (Issue #460)
 
 * Fixed bug in Image class resize function.  (Issue #461)
+
+* Sped up the OpticalPSF calculations significantly.  (Issue #466)